Fixed a boost-related crash.
In rime::Deployer::JoinMaintenanceThread(), the call
to boost::thread::join() may cause application to crash
if the thread is not joinable (i.e. not-a-thread, or thread
terminated).

However, it seems that this situtaion only occur with the
recent version of boost (without C++11, with multithread,
with BOOST_THREAD_TRROW_IF_PRECONDITION_NOT_SATISFIED
marco defined by default).