CMakeLists.txt: Ensure path in pkgconfig file to be absolute
This fixes the regression introduced in the last commit.
The ${CMAKE_INSTALL_*DIR} variables are not absolute and we
must prepend them with ${prefix} or ${exec_prefix} to make it
fully working.

There are also ${CMAKE_INSTALL_FILL_*DIR} variables but we are
not using them for now since librime is still using custom
${LIB_INSTALL_DIR} and ${BIN_INSTALL_DIR}.