feat(api): get_state_label returns the state label for UI display
enable the frontend to display a message when an arbitrary option defined in the
schema is updated.

Closes #447

the option name eg. ascii_mode is passed to the notification handler function,
there, get_state_label gives IME frontend the access to the human readable label
for the current state of the option defined in *.schema:/switches/@*/states.

when selection is changed in a radio group, more than one option in the group is
updated so it only returns a non-empty label while querying the selected option.