refactor(dict): speed up `EncodePhrase` when building a table
`EncodePhrase` is one of the most time-consuming methods to build a table with its sub method `CreateEntry`. Since C++ 11 (and we are using C++ 14 standard), `std::sto*` and `std::to_string` methods are available, and they basically play the same role of `boost::lexical_cast`, and have better performance. So I replace involved `boost::lexical_cast` methods with them.

Also I replace `push_back` with `emplace_back` for the involved variables of vector type, the latter is a faster method to append a element to a vector; and replace involved the set type of the involved variables with unordered_set type to get better performance, since they seem to have no requirement for order.

This may save about 2 seconds averagely base on the last commit (total time: ~15s -> ~13s).