feat(table_translator): Allow sorting completions
Add a new boolean option `sort_completions`. When enabled (requiring
`enable_completion` also to be enabled), completion candidates will be
sorted according to the weights specified in the dictionary.

This is done by automatically generating kCompletion spellings in the
prism, as if they were generated from spelling algebra written by the
user.

Previously, the order was arbitrary (but more-or-less in alphabet
order).